feat: MET-1683 refactor governance votes
