+++ src/commons/utils/constants.ts
@@ -302,6 +302,7 @@ export enum POOLS_ACTION_TYPE {
   INFO_ACTION = "INFO_ACTION"
 }
 
-export enum POOLS_VOTE_TYPE {
-  STAKING_POOL_KEY_HASH = "STAKING_POOL_KEY_HASH"
+export enum VOTE_TYPE {
+  STAKING_POOL_KEY_HASH = "STAKING_POOL_KEY_HASH",
+  DREP_KEY_HASH = "DREP_KEY_HASH"
 }
