+++ src/components/commons/CardGovernanceVotes/styles.ts
@@ -1,11 +1,13 @@
 import { Box, styled } from "@mui/material";
 
-export const CardGovernanceVote = styled(Box)`
-  padding: 20px;
-  border-radius: 8px;
-  cursor: pointer;
+import Card from "../Card";
+
+export const CardGovernanceVote = styled(Card)`
+  width: 100%;
+  height: 100%;
+  background: ${({ theme }) => (theme.isDark ? theme.palette.secondary[100] : theme.palette.secondary[700])};
   box-shadow: 2px 2px 10px 0px #43465633;
-  background: ${({ theme }) => (theme.isDark ? theme.palette.secondary[100] : "")};
+  cursor: pointer;
   &:hover {
     box-shadow: ${({ theme }) =>
       theme.isDark ? ` 2px 2px 10px 0px ${theme.palette.secondary[100]}` : theme.shadow.cardHover};
