+++ src/locales/en/translation.json
@@ -506,7 +506,8 @@
   "common.poolhash": "Pool hash",
   "common.poolHomepage": "Homepage",
   "common.poolDescription": "Description",
-  "common.poolName": "Pool Name ",
+  "common.poolName": "Pool Name",
+  "common.overall": "Overall Vote",
   "glossary.epoch": "Epoch",
   "glossary.upcomingEpoch": "Upcoming Epoch",
   "common.rewardAccount": "Reward Account",
